requirements.txt
groq
pydantic
python-dotenv
tqdm
